+++ __tests__/OmnichannelChatSDK.spec.ts
@@ -2499,7 +2499,7 @@ describe('Omnichannel Chat SDK', () => {
 
             const dummyConfig = {
                 LiveWSAndLiveChatEngJoin: {
-                    msdyn_postconversationsurveyenable: false,
+                    msdyn_postconversationsurveyenable: "false",
                     msfp_sourcesurveyidentifier: "",
                     postConversationSurveyOwnerId: ""
                 }
@@ -2533,7 +2533,7 @@ describe('Omnichannel Chat SDK', () => {
 
             const dummyConfig = {
                 LiveWSAndLiveChatEngJoin: {
-                    msdyn_postconversationsurveyenable: true,
+                    msdyn_postconversationsurveyenable: "true",
                     msfp_sourcesurveyidentifier: "",
                     postConversationSurveyOwnerId: ""
                 },
@@ -2572,7 +2572,7 @@ describe('Omnichannel Chat SDK', () => {
 
             const dummyConfig = {
                 LiveWSAndLiveChatEngJoin: {
-                    msdyn_postconversationsurveyenable: true,
+                    msdyn_postconversationsurveyenable: "true",
                     msfp_sourcesurveyidentifier: "",
                     postConversationSurveyOwnerId: ""
                 },
@@ -2614,7 +2614,7 @@ describe('Omnichannel Chat SDK', () => {
 
             const dummyConfig = {
                 LiveWSAndLiveChatEngJoin: {
-                    msdyn_postconversationsurveyenable: true,
+                    msdyn_postconversationsurveyenable: "true",
                     msfp_sourcesurveyidentifier: "",
                     postConversationSurveyOwnerId: ""
                 },
@@ -2664,7 +2664,7 @@ describe('Omnichannel Chat SDK', () => {
 
             const dummyConfig = {
                 LiveWSAndLiveChatEngJoin: {
-                    msdyn_postconversationsurveyenable: true,
+                    msdyn_postconversationsurveyenable: "true",
                     msfp_sourcesurveyidentifier: "",
                     postConversationSurveyOwnerId: ""
                 },
@@ -2696,6 +2696,65 @@ describe('Omnichannel Chat SDK', () => {
                 expect(chatSDK.OCClient.getLWIDetails).toHaveBeenCalledTimes(1);
                 expect(chatSDK.OCClient.getSurveyInviteLink).toHaveBeenCalledTimes(1);
                 expect(postChatContext.participantJoined).toBeFalsy();
+            } catch (ex) {
+                throw("Should not throw error. " + ex);
+            }
+        });
+
+        it('ChatSDK.getPostChatSurveyContext() should resolve if bot survey is being used', async () => {
+            const chatSDKConfig = {
+                telemetry: {
+                    disable: true
+                },
+                chatReconnect: {
+                    disable: false,
+                }
+            };
+
+            const dummyConfig = {
+                LiveWSAndLiveChatEngJoin: {
+                    msdyn_postconversationsurveyenable: "true",
+                    msfp_sourcesurveyidentifier: "",
+                    msfp_botsourcesurveyidentifier: "1",
+                    postConversationSurveyOwnerId: "",
+                    postConversationBotSurveyOwnerId: "2"
+                },
+                ChatWidgetLanguage: {
+                    msdyn_localeid: "1033"
+                }
+            };
+
+            const chatSDK = new OmnichannelChatSDK(omnichannelConfig, chatSDKConfig);
+
+            await chatSDK.initialize();
+
+            chatSDK.liveChatConfig = dummyConfig;
+            jest.spyOn(chatSDK.OCClient, 'getLWIDetails').mockResolvedValue({
+                ConversationId: "dummy",
+                CanRenderPostChat: "True",
+                ParticipantType: "Bot"
+            });
+            jest.spyOn(chatSDK.OCClient, 'getSurveyInviteLink').mockResolvedValue({
+                inviteList: [
+                    {
+                        invitationLink: "dummy"
+                    }
+                ],
+                formsProLocaleCode: "en-us"
+            });
+            jest.spyOn(console, 'error');
+
+            try {
+                const postChatContext = await chatSDK.getPostChatSurveyContext();
+                expect(chatSDK.OCClient.getLWIDetails).toHaveBeenCalledTimes(1);
+                expect(chatSDK.OCClient.getSurveyInviteLink).toHaveBeenCalledWith("2", {
+                    "FormId": "1",
+                    "ConversationId": "dummy",
+                    "OCLocaleCode": "en-us"
+                },
+                expect.any(Object));
+                expect(postChatContext.participantJoined).toBeTruthy();
+                expect(postChatContext.participantType).toBe("Bot");
             } catch (ex) {
                 throw("Should not throw error. " + ex);
             }